Refactor line to separate module.
